+++ CHANGELOG.md
@@ -1,11 +1,24 @@
 # Change Log
 
+## 0.3.3 ()
+
+Fixed
+1. Fixed bug that prevented custom sqlite paths from being recognized as a command on Windows.
+2. Fixed bug that sometimes truncated the error message.
+
+Added
+1. Command: ```SQLite: Show output```, to show the extension's output channel.
+
+Other
+1. Updated SQLite3 binaries to the version 3.26.0
+
 ## 0.3.2 (24 Nov 2018)
 
 Fixed
 1. Fixed bug that prevented the extension from recognizing executable paths set in the setting ```sqlite.sqlite3```.
 
 ## 0.3.1 (5 Sep 2018)
+
 Added
 1. Autocompletion for sqlite keywords, table names, column names.
 2. Added ```.db3```, ```.sdb```, ```.s3db``` as recognized extensions for an sqlite3 database (thanks to [LokiSharp](https://github.com/LokiSharp))
@@ -15,6 +28,7 @@ Fixed
 1. Fixed issue that prevented the correct execution of queries on tables with the same name as sqlite keywords (e.g Order, Select, Table...)
 
 ## 0.3.0 (2 Sep 2018)
+
 Added
 1. Export result to ```html```.
 2. Export multiple results to ```csv```, ```html``` or ```json```.
@@ -33,10 +47,12 @@ Other
 
 
 ## 0.2.3 (12 Jul 2018)
+
 Added
 1. Setting: ```sqlite.outputBuffer```, query output buffer in bytes.
 
 ## 0.2.2 (12 Jul 2018)
+
 Added
 1. Pagination, tables are now displayed in pages, each page with a set number of records.
 2. Setting: ```sqlite.recordsPerPage```, to set the number of records per page.
@@ -49,10 +65,12 @@ Fixed
 2. Fixed a bug that prevented the correct execution on Windows of queries composed of multiple statements.
 
 ## 0.2.1 (11 Jul 2018)
+
 Fixed
 1. Fixed bug that prevented the query result from showing on Windows.
 
 ## 0.2.0 (11 Jul 2018)
+
 Added
 1. Export query result to Json.
 2. Export query result to Csv.
@@ -65,10 +83,12 @@ Fixed
 1. Explorer: Show Table can now display more records.
 
 ## 0.1.2 (5 Jul 2018)
+
 Fixed
 1.  Fixed bug that prevented the database from showing in the explorer if it was previously closed.
 
 ## 0.1.1 (5 Jul 2018)
+
 Added
 1. New setting: ```sqlite3.logLevel```, to set the logging level of the output channel.
 2. New setting: ```sqlite3.showTableLimit```, to limit the records displayed when showing a table using Show Table (in SQLite Explorer).
@@ -78,6 +98,7 @@ Changed
 2. SQLite output channel now displays time and log level.
 
 ## 0.1.0 (29 Jun 2018)
+
 Added
 1. Added new command: Quick Query, execute a query directly without creating a new sql document.
 2. Tables in the explorer are now ordered alphabetically.
@@ -89,10 +110,12 @@ Fixed
 2. Comments inside a comment in an sql document are now parsed correctely.
 
 ## 0.0.4 (26 Jun 2018)
+
 Fixed
 1. unicode characters are now shown correctely
 
 ## 0.0.3 (26 Jun 2018)
+
 Fixed
 1. fixed a bug that prevented the correct parsing of the query result when it contained an escaped quote
 2. fixed a bug that crashed the extension when the query result contained a unicode character inside the string
@@ -101,8 +124,10 @@ Other
 1. improved quality of the icon
 
 ## 0.0.2 (25 Jun 2018)
+
 Fixed
 1. fixed a bug that showed unicode characters as octets.
 
 ## 0.0.1 (23 Jun 2018)
- Initial release
+
+Initial release